split virtual list
